# openapi.yml
openapi: 3.0.3
info:
  title: VoidRun API
  description: |
    VoidRun API provides comprehensive management of virtual machines (sandboxes), 
    file system operations, execution environments, and organizational resources.

    All endpoints except `/api/register` require the `X-API-Key` header for authentication.
  version: 1.0.0
  contact:
    name: VoidRun Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:33944/api
    description: Local development server
  - url: https://api.void-run.com/api
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for authentication (format: hf_...)"
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          description: Response data (structure varies by endpoint)
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        details:
          type: string
          example: Additional error details

    # Registration
...
paths:
...
  /sandboxes/{id}/files/upload:
    post:
      tags:
        - File System
      summary: Upload file
      description: Upload a file to the sandbox
      operationId: uploadFile
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /root/script.sh
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...